fix: 뷰, 뷰모델 네이밍
